third_party/blink/web_tests/svg/dynamic-updates/SVGFECompositeElement-dom-k1-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'k1' attribute of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.getAttribute('k1') is "1.9"
PASS overComposite1.getAttribute('k1') is ".5"
PASS successfullyParsed is true

TEST COMPLETE
